requirements.txt
streamlit
streamlit-chat
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client 
pydrive 
gspread
Levenshtein
scikit-learn